apps/api/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
SQLAlchemy>=2.0.0,<3.0.0
psycopg[binary]>=3.1.0,<4.0.0
pydantic-settings>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
openai>=1.0.0,<2.0.0
